DRF/requirements.txt
gunicorn==20.1.0
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
hiredis==2.0.0
idna==3.3
multidict==6.0.2
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
psycopg2-binary>=2.8
django-cors-headers
